+++ firestore_annotations/lib/firestore_annotations.dart
@@ -1,11 +1,13 @@
 /* Firestore Serializable Annotations */
+library firestore_annotations;
 
 class FirestoreDocument {
   final bool hasSelfRef;
-  final bool flutterFormHelper;
+  final bool nullable;
+
   const FirestoreDocument({
+    this.nullable = false,
     this.hasSelfRef = true,
-    this.flutterFormHelper = false,
   });
 }
 
@@ -15,13 +17,11 @@ class FirestoreAttribute {
   final bool nullable;
   final String alias;
   final dynamic defaultValue;
-  final String flutterValidatorMessage;
 
   const FirestoreAttribute({
     this.ignore = false,
     this.nullable = true,
     this.alias = null,
     this.defaultValue = null,
-    this.flutterValidatorMessage = 'Couldn\'t parse input!',
   });
 }