+++ firestore_functions/lib/firestore_functions.dart
@@ -1,0 +1,45 @@
+library firestore_functions;
+
+export 'src/functions_impl.dart' show functions, FirebaseFunctions;
+
+export 'package:firestore_api/firestore_api.dart';
+export 'package:firestore_node/firestore_node.dart';
+
+export 'package:firebase_functions_interop/firebase_functions_interop.dart'
+    show
+        HttpsFunctions,
+        CallableContext,
+        HttpsError,
+        DataEventHandler,
+        ChangeEventHandler,
+        Config,
+        Change,
+        EventContext,
+        DatabaseFunctions,
+        RefBuilder,
+        PubsubFunctions,
+        TopicBuilder,
+        ScheduleBuilder,
+        Message,
+        StorageFunctions,
+        BucketBuilder,
+        ObjectBuilder,
+        ObjectMetadata,
+        CustomerEncryption,
+        AuthFunctions,
+        UserBuilder,
+        UserRecord,
+        CloudFunction,
+        HttpsFunction,
+        EventAuthInfo,
+        RuntimeOptions;
+export 'package:firebase_functions_interop/src/express.dart';
+
+export 'package:firebase_admin_interop/src/admin.dart';
+export 'package:firebase_admin_interop/src/app.dart';
+export 'package:firebase_admin_interop/src/auth.dart' hide UserRecord;
+export 'package:firebase_admin_interop/src/bindings.dart'
+    show AppOptions, SetOptions, FirestoreSettings;
+export 'package:firebase_admin_interop/src/messaging.dart';
+
+export 'package:node_io/node_io.dart' show HttpRequest, HttpResponse;