+++ firestore_node/lib/firestore_node.dart
@@ -1,0 +1,3 @@
+library firestore_node;
+
+export 'src/node_impl.dart' show FirestoreImpl, DocumentSnapshotImpl;