+++ firestore_serializable/lib/src/generator.dart
@@ -1,8 +1,7 @@
 import 'package:build/build.dart';
 import 'package:firestore_serializable/src/annotation_helper.dart';
-import 'package:firestore_serializable/src/form_helper.dart';
-import 'package:firestore_serializable/src/map_helper.dart';
-import 'package:firestore_serializable/src/snapshot_helper.dart';
+import 'package:firestore_serializable/src/serialize_helper.dart';
+import 'package:firestore_serializable/src/deserialize_helper.dart';
 import 'package:source_gen/source_gen.dart';
 import 'package:analyzer/dart/element/element.dart';
 import 'package:build/src/builder/build_step.dart';
@@ -38,30 +37,18 @@ class _Generator {
     final accessibleFields = <FieldElement>[];
 
     for (var el in element.fields) {
-      // TODO find more elegant solution
-      // TODO ignore getter, if no setter with the same name is available
-      if (!el.isPublic) {
-        //throw 'Error';
-      } else if (el.name == 'selfRef') {
-        //throw 'Error';
-      } else if (el.getter == null) {
-        //throw 'Error';
-      } else {
+      if (el.isPublic && el.name != 'selfRef') {
         accessibleFields.add(el);
       }
     }
 
-    SnapshotHelper snapshotHelper = SnapshotHelper(className);
+    DeserializeHelper deserializeHelper = DeserializeHelper(className);
 
-    yield* snapshotHelper.createFromSnapshot(
-        accessibleFields, annotationHelper.hasSelfRef);
-    yield* snapshotHelper.createFromMap(accessibleFields, className);
+    yield* deserializeHelper.createFromSnapshot(accessibleFields,
+        annotationHelper.hasSelfRef, annotationHelper.nullable);
+    yield* deserializeHelper.createFromMap(
+        accessibleFields, className, annotationHelper.nullable);
 
-    yield* MapHelper(className).createToMap(accessibleFields);
-
-    if (annotationHelper.flutterFormHelper) {
-      yield* FormHelper(className)
-          .createHelperExtension(accessibleFields, className);
-    }
+    yield* SerializeHelper(className).createToMap(accessibleFields);
   }
 }